Move SalClient into its own module.
