+++ sal_python_pkg/sal/__init__.py
@@ -1,2 +1,3 @@
+from sal.client import SalClient
 from sal.utils import *
 from sal.version import __version__
