+++ sal_python_pkg/sal/exceptions.py
@@ -1,0 +1,2 @@
+class SalClientError(Exception):
+    pass
